requirements.txt
Flask
Django==3.1.6
requests
jaeger-client
opentracing-instrumentation
tornado==5.1.1  # tornado==6.0.4
